+++ Server/Server.Map/Map/EpPathFinding/Heuristic.cs
@@ -36,8 +36,6 @@ An Interface for the Heuristic Function Class.
 
 */
 using System;
-using System.Collections.Generic;
-using System.Collections;
 
 namespace EpPathFinding
 {
@@ -58,14 +56,14 @@ namespace EpPathFinding
 
         public static float Euclidean(int iDx, int iDy)
         {
-            float tFdx = (float)iDx;
-            float tFdy = (float)iDy;
-            return (float)Math.Sqrt((double)(tFdx * tFdx + tFdy * tFdy));
+            float tFdx = iDx;
+            float tFdy = iDy;
+            return (float)Math.Sqrt(tFdx * tFdx + tFdy * tFdy);
         }
 
         public static float Chebyshev(int iDx, int iDy)
         {
-            return (float)Math.Max(iDx, iDy);
+            return Math.Max(iDx, iDy);
         }
     }
 }
